refactor(frontend): split AgentReviewStep into sub-components and a hook
AgentReviewStep had grown past the ~200 line guideline. Extract derived state into useAgentReviewStep and the hero, summary card, and footer into local sub-components. Fold the hero accent (pulse/gradient) directly into each getHeroContent branch instead of inferring it from Tailwind class substrings, removing the fragile getHeroAccent lookup.
